+++ src/main/kotlin/lotto/display/BonusNumberDisplay.kt
@@ -1,0 +1,7 @@
+package lotto.display
+
+class BonusNumberDisplay {
+    fun show(bonusNumber: Int) {
+        print("Bonus Number entered: $bonusNumber")
+    }
+}